+++ src/formats/html.ts
@@ -1,0 +1,7 @@
+import { Segment } from "../types"
+
+export const parseHTML = (data: string): Array<Segment> => {
+    const outSegments = []
+    console.log("parseHTML")
+    return outSegments
+}
