+++ src/formats/json.ts
@@ -1,0 +1,6 @@
+import { Segment } from "../types"
+
+export type JSONTranscript = {
+    version: string
+    segments: Array<Segment>
+}
