+++ src/formats/vtt.ts
@@ -1,0 +1,7 @@
+import { Segment } from "../types"
+
+export const parseVTT = (data: string): Array<Segment> => {
+    const outSegments = []
+    console.log("parseVTT")
+    return outSegments
+}
